requirements.txt
mysql-connector-python~=8.0.29
pytz~=2022.1
git+https://github.com/Rapptz/discord.py
python-dotenv
